change db again

+++ nasu/database/migrations/2025_05_28_214558_create_room_items_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->foreignId('room_id')->constrained();
             $table->foreignId('user_furniture_id')->constrained('user_furniture');
-            $table->decimal('x_position', 5, 2); 
-            $table->decimal('y_position', 5, 2);
+            $table->integer('x_position'); 
+            $table->integer('y_position');
             $table->integer('rotation')->default(0);
             $table->timestamps();
         });
